Example of Robyn 3.9 Code Deployment
This version of Robyn is leveraging a calibration file based on MTA and GeoLift test results.
Cores are also set to leverage MacPro M1, which has ten cores but I have set it to 8 because trials are only 5 and iterations are set to 10k. If the trials increase, then I would set it to 10.